multiple region rates now tentatively working

+++ src/nrnpython/grids.cpp
@@ -312,8 +312,8 @@ extern "C" void set_grid_concentrations(int grid_list_index, int index_in_list, 
     */
     /* TODO: note that these will need updating anytime the structure of the model changes... look at the structure change count at each advance and trigger a callback to regenerate if necessary */
     Grid_node* g;
-    Py_ssize_t i;
-    Py_ssize_t n = PyList_Size(grid_indices);
+    ssize_t i;
+    ssize_t n = (ssize_t)PyList_Size(grid_indices);
 
     /* Find the Grid Object */
     g = Parallel_grids[grid_list_index];
@@ -347,8 +347,8 @@ extern "C" void set_grid_currents(int grid_list_index, int index_in_list, PyObje
     */
     /* TODO: note that these will need updating anytime the structure of the model changes... look at the structure change count at each advance and trigger a callback to regenerate if necessary */
     Grid_node* g;
-    Py_ssize_t i;
-    Py_ssize_t n = PyList_Size(grid_indices);
+    ssize_t i;
+    ssize_t n = (ssize_t)PyList_Size(grid_indices);
     long* dests;
 
     /* Find the Grid Object */
@@ -574,7 +574,7 @@ int ECS_Grid_node::dg_adi()
 
 void ECS_Grid_node::scatter_grid_concentrations()
 {
-    Py_ssize_t i, n;
+    ssize_t i, n;
     Concentration_Pair* cp;
     double* my_states;
